% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_heb.R
\name{str_heb}
\alias{str_heb}
\title{Add a Unicode RTL-Isolating character \verb{\\u2067} to a character vector}
\usage{
str_heb(string = NULL)
}
\arguments{
\item{string}{A character vector. The default is \code{NULL} in case there is a need to only use the Unicode character.}
}
\value{
A character vector.
}
\description{
Add a Unicode RTL-Isolating character \verb{\\u2067} to a character vector
}
\examples{
x <- "סימני ניקוד בסוף המשפט."
str_heb(x)
}
